requirement.txt
Dataset from kaggle
pretrained model - resnet18

environment requirements

Package           Version
----------------- -----------
contourpy         1.2.0
cycler            0.12.1
filelock          3.13.3
fonttools         4.50.0
fsspec            2024.3.1
future            1.0.0
Jinja2            3.1.3
kiwisolver        1.4.5
MarkupSafe        2.1.5
matplotlib        3.8.3
mpmath            1.3.0
networkx          3.2.1
numpy             1.26.4
packaging         24.0
pillow            10.2.0
pip               24.0
pyparsing         3.1.2
python-dateutil   2.9.0.post0
six               1.16.0
sympy             1.12
torch             2.2.1
torchvision       0.17.1
typing_extensions 4.10.0
